requirements.txt
# Core libraries used in the project
numpy>=1.21
Pillow>=8.0

# PyTorch and torchvision (install torch per your CUDA / CPU environment)
torch>=2.0.0
torchvision>=0.15.0

# Model & HPO
timm>=0.9.0
optuna>=3.0.0